+++ scheduler/crud/flyte_scheduler.go
@@ -1,4 +1,4 @@
-package flytescheduler
+package crud
 
 import (
 	"context"
